Backport 2abf4cb43e84dda883a38abc6bcbac5aba3d7ce8
