+++ src/hotspot/share/gc/shenandoah/shenandoahDegeneratedGC.hpp
@@ -66,6 +66,8 @@ private:
   void upgrade_to_full();
 
   const char* degen_event_message(ShenandoahDegenPoint point) const;
+
+  bool has_in_place_promotions(const ShenandoahHeap* heap) const;
 };
 
 #endif // SHARE_GC_SHENANDOAH_SHENANDOAHDEGENERATEDGC_HPP
